requirements.txt
streamlit
numpy
tensorflow
opencv-python
pandas
pillow
scikit-learn
joblib